Absolutely destroy code in code review.
