--- Core/ECS/CameraEntity.cs
@@ -1,16 +1,0 @@
-﻿using UntitledEngine.Core.Components;
-using UntitledEngine.Core.Entities;
-
-namespace UntitledEngine.Core.ECS;
-
-public class CameraEntity : Entity
-{
-    public CameraEntity()
-    {
-        // Add Camera component to the GameObject
-        Name = "Camera";
-        AddTag("Camera"); // Add Camera tag
-        var camera = new Camera();
-        AddComponent(camera);
-    }
-}
