+++ Core/ECS/Entity.cs
@@ -16,11 +16,8 @@ public class Entity
     public Transform? Transform => _components.OfType<Transform>().FirstOrDefault();
 
     // Constructor to initialise the GameObject
-    public Entity()
-    {
-        AddComponent(new Transform());
-    }
-    
+    public Entity() => AddComponent(new Transform());
+
     public void Destroy()
     {
         // If PhysicsBody found, erase its body before deleting component
@@ -64,17 +61,11 @@ public class Entity
     // --- TAGS ---
     
     // Add tag to entity
-    public void AddTag(string tag)
-    {
-        _tags.Add(tag);
-    }
-    
+    public void AddTag(string tag) => _tags.Add(tag);
+
     // Remove tag from entity
-    public void RemoveTag(string tag)
-    {
-        _tags.Remove(tag);
-    }
-    
+    public void RemoveTag(string tag) => _tags.Remove(tag);
+
     // Check if entity has tag
     public bool HasTag(string tag)
     {
