+++ Core/Renderer/Renderer.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace UntitledEngine.Core.Renderer;
+
+internal static class Renderer
+{
+    public static void OnLoad()
+    {
+
+    }
+
+    public static void OnRender(double deltaTime)
+    {
+
+    }
+}
