+++ UntitledEngine.csproj.user
@@ -1,0 +1,21 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="Current" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup />
+  <ItemGroup>
+    <Compile Update="Core\Components\Camera.cs">
+      <SubType>Component</SubType>
+    </Compile>
+    <Compile Update="Core\Components\MeshRenderer.cs">
+      <SubType>Component</SubType>
+    </Compile>
+    <Compile Update="Core\Components\PhysicsBody.cs">
+      <SubType>Component</SubType>
+    </Compile>
+    <Compile Update="Core\Components\PlayerController.cs">
+      <SubType>Component</SubType>
+    </Compile>
+    <Compile Update="Core\Components\Transform.cs">
+      <SubType>Component</SubType>
+    </Compile>
+  </ItemGroup>
+</Project>
