+++ imgui.ini
@@ -23,3 +23,8 @@ Pos=9,41
 Size=283,623
 Collapsed=0
 
+[Window][Inspector]
+Pos=538,73
+Size=328,408
+Collapsed=0
+
